requirements.txt
asgiref==3.11.0
Django==5.2.8
psycopg2-binary==2.9.11
sqlparse==0.5.4
django-allauth==65.0.2
python-dotenv==1.0.1
channels==4.0.0
channels-redis==4.2.0
PyJWT==2.10.1
requests
pillow
channels==4.0.0
channels-redis==4.2.0
